Update observability script
